Resolve broken test cases
The test which used the real PCRE lib to confirm that generated strings
actually conformed to the regex found a number of edge cases. This
update resolves a lot of them.

We also fix a number of issues around building the application as root,
which can result in generated files not having a known user associated
with them, and hence resulting in permissions errors.