Handle escaping nonalphanumeric characters properly
We weren't parsing certain nonalphanumeric characters in such a way that
they would be rendered correctly.

Also introduced the `ormolu` autoformatter in the middle of work, which
has meant that the diff for this is unnecessarily messy. Sorry but it
was easier to commit the reformatted files than to unpick the changes
for a one-person project. 😕