Produce example strings containing back references
One of the requested features for regexicon.com was handling of
back references in regex, which the DFA library did not handle. This
update is an initial implementation of that.

Back references in PCRE are an awkward thing to parse using Parsec as
the common `\1` notation overlaps with the notation for octal
characters. The implementation here uses a second transformation phase
to reprocess the string once all the subsequences are known

I've certain concerns about this as it means that the Quantifiable type
includes three constructors, `AmbiguousNumberSequence`,
`BackslashSequence` and `BackReference`, which should not coexist. The
obvious possibility is to duplicate the types, so that there is an
ambiguous and non-ambiguous version, but that will be very unwieldy to
maintain. Phantom types may at least allow us to know that we're in one
state or the other, but I'm also going to investigate if a monad
transformer might allow this to work on the fly, but it's possible that
this may not be something that's compatible with how Parsec works 🤔