Handle `\x` (NonprintingHexZero)
`\x` is a weird corner case in the PCRE grammar. Without a following hex
digit it should be interpreted as a literal `\NUL` character, if
followed by a `{` any following hex digits will be understood to
represent one or more code points.

However `\x` is a quantifiable regex character, you should be able to
define repeated strings using `*` or `+`. However PCRE will _not_
quantify using the min max quantifier (ie `\x{1,3}`) as the grammar can
overlap with the previous forms.

This pull request adds handling for this form in the parsing and
construction code for `NonprintingHexZero`.